+++ gems/aws-sdk-translate/lib/aws-sdk-translate/types.rb
@@ -492,26 +492,26 @@ module Aws::Translate
     #   as input. You can specify one of the following multipurpose internet
     #   mail extension (MIME) types:
     #
-    #   * `text/html`: The input data consists of one or more HTML files.
+    #   * `text/html`\: The input data consists of one or more HTML files.
     #     Amazon Translate translates only the text that resides in the
     #     `html` element in each file.
     #
-    #   * `text/plain`: The input data consists of one or more unformatted
+    #   * `text/plain`\: The input data consists of one or more unformatted
     #     text files. Amazon Translate translates every character in this
     #     type of input.
     #
-    #   * `application/vnd.openxmlformats-officedocument.wordprocessingml.document`:
+    #   * `application/vnd.openxmlformats-officedocument.wordprocessingml.document`\:
     #     The input data consists of one or more Word documents (.docx).
     #
-    #   * `application/vnd.openxmlformats-officedocument.presentationml.presentation`:
+    #   * `application/vnd.openxmlformats-officedocument.presentationml.presentation`\:
     #     The input data consists of one or more PowerPoint Presentation
     #     files (.pptx).
     #
-    #   * `application/vnd.openxmlformats-officedocument.spreadsheetml.sheet`:
+    #   * `application/vnd.openxmlformats-officedocument.spreadsheetml.sheet`\:
     #     The input data consists of one or more Excel Workbook files
     #     (.xlsx).
     #
-    #   * `application/x-xliff+xml`: The input data consists of one or more
+    #   * `application/x-xliff+xml`\: The input data consists of one or more
     #     XML Localization Interchange File Format (XLIFF) files (.xlf).
     #     Amazon Translate supports only XLIFF version 1.2.
     #